+++ trixnity-client/trixnity-client-repository-room/src/commonMain/kotlin/net/folivo/trixnity/client/store/repository/room/RoomOlmForgetFallbackKeyAfterRepository.kt
@@ -1,0 +1,57 @@
+package net.folivo.trixnity.client.store.repository.room
+
+import androidx.room.Dao
+import androidx.room.Entity
+import androidx.room.Insert
+import androidx.room.OnConflictStrategy
+import androidx.room.PrimaryKey
+import androidx.room.Query
+import kotlinx.datetime.Instant
+import net.folivo.trixnity.client.store.repository.OlmForgetFallbackKeyAfterRepository
+
+@Entity(tableName = "OlmForgetFallbackKeyAfter")
+data class RoomOlmForgetFallbackKeyAfter(
+    @PrimaryKey val id: Long,
+    val instant: Instant,
+)
+
+@Dao
+interface OlmForgetFallbackKeyAfterDao {
+    @Query("SELECT * FROM OlmForgetFallbackKeyAfter WHERE id = :id LIMIT 1")
+    suspend fun get(id: Long): RoomOlmForgetFallbackKeyAfter?
+
+    @Insert(onConflict = OnConflictStrategy.REPLACE)
+    suspend fun insert(entity: RoomOlmForgetFallbackKeyAfter)
+
+    @Query("DELETE FROM OlmForgetFallbackKeyAfter WHERE id = :id")
+    suspend fun delete(id: Long)
+
+    @Query("DELETE FROM OlmForgetFallbackKeyAfter")
+    suspend fun deleteAll()
+}
+
+internal class RoomOlmForgetFallbackKeyAfterRepository(
+    db: TrixnityRoomDatabase,
+) : OlmForgetFallbackKeyAfterRepository {
+    private val dao = db.olmForgetFallbackKeyAfter()
+
+    override suspend fun get(key: Long): Instant? =
+        dao.get(key)?.instant
+
+    override suspend fun save(key: Long, value: Instant) {
+        dao.insert(
+            RoomOlmForgetFallbackKeyAfter(
+                id = key,
+                instant = value,
+            )
+        )
+    }
+
+    override suspend fun delete(key: Long) {
+        dao.delete(key)
+    }
+
+    override suspend fun deleteAll() {
+        dao.deleteAll()
+    }
+}